+++ backend/de.metas.adempiere.adempiere/base/src/main/java/de/metas/scheduler/housekeeping/ResetSchedulerState.java
@@ -1,10 +1,8 @@
-package org.adempiere.scheduler.housekeeping.spi.impl;
-
 /*
  * #%L
  * de.metas.adempiere.adempiere.base
  * %%
- * Copyright (C) 2015 metas GmbH
+ * Copyright (C) 2022 metas GmbH
  * %%
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -21,6 +19,8 @@ package org.adempiere.scheduler.housekeeping.spi.impl;
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
+
+package de.metas.scheduler.housekeeping;
 
 import org.adempiere.ad.housekeeping.spi.IStartupHouseKeepingTask;
 import org.adempiere.ad.trx.api.ITrx;
