+++ backend/de.metas.adempiere.adempiere/base/src/test/java/de/metas/scheduler/SchedulerDaoTest.java
@@ -2,7 +2,7 @@
  * #%L
  * de.metas.adempiere.adempiere.base
  * %%
- * Copyright (C) 2021 metas GmbH
+ * Copyright (C) 2022 metas GmbH
  * %%
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -20,9 +20,10 @@
  * #L%
  */
 
-package org.adempiere.scheduler;
+package de.metas.scheduler;
 
 import de.metas.process.AdProcessId;
+import de.metas.scheduler.SchedulerDao;
 import org.adempiere.exceptions.AdempiereException;
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.adempiere.test.AdempiereTestHelper;
